fix: 에러 핸들링 추가

+++ src/main/java/com/project/demo/wishlist/controller/WishlistController.java
@@ -3,6 +3,7 @@ package com.project.demo.wishlist.controller;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.project.demo.auth.domain.User;
+import com.project.demo.exceptions.NotAllowedUser;
 import com.project.demo.wishlist.dto.RequestCreateWishlist;
 import com.project.demo.wishlist.dto.ResponseIsExist;
 import com.project.demo.wishlist.dto.ResponseWishlist;
@@ -26,28 +27,41 @@ public class WishlistController {
     }
 
     @GetMapping("wishlist")
-    public ResponseWishlist getWishlist(HttpSession session) {
+    public ResponseWishlist getWishlist(HttpSession session) throws NotAllowedUser {
         User user = (User)session.getAttribute("user");
+        if(user == null) {
+            throw new NotAllowedUser();
+        }
         ResponseWishlist res = new ResponseWishlist(this.wishlistService.getWishlist(user.getId()));
         return res;
     }
 
     @PostMapping("wishlist")
-    public void createWishlist(HttpSession session, @RequestBody RequestCreateWishlist req) {
+    public void createWishlist(HttpSession session, @RequestBody RequestCreateWishlist req) throws NotAllowedUser{
         User user = (User)session.getAttribute("user");
+        if(user == null) {
+            throw new NotAllowedUser();
+        }
         this.wishlistService.appendWishlist(user.getId(), req.getProductId());
     }
 
     @GetMapping("wishlist/{productId}")
-    public ResponseIsExist CheckDuplication(HttpSession session, @PathVariable("productId") int productId){
+    public ResponseIsExist CheckDuplication(HttpSession session, @PathVariable("productId") int productId) throws NotAllowedUser{
         User user = (User)session.getAttribute("user");
+        if(user == null) {
+            throw new NotAllowedUser();
+        }
+
         boolean isExist = this.wishlistService.isExistWishlist(user.getId(), productId);
         return new ResponseIsExist(isExist);
     }
 
     @DeleteMapping("wishlist/{productId}")
-    public void deleteWishlist(HttpSession session, @PathVariable("productId") int productId) {
+    public void deleteWishlist(HttpSession session, @PathVariable("productId") int productId) throws NotAllowedUser{
         User user = (User)session.getAttribute("user");
+        if(user == null) {
+            throw new NotAllowedUser();
+        }
         this.wishlistService.deleteWishlist(user.getId(), productId);
     }
 }
